fixed part of warnings

+++ src/plugins/intel_gpu/src/graph/include/arg_max_min_inst.h
@@ -20,7 +20,7 @@ public:
     program_node& input() const { return get_dependency(0); }
 
     uint32_t get_output_nums() const {
-        return (get_primitive()->input_size() == 3 ? 2 : get_primitive()->output_size());
+        return static_cast<uint32_t>(get_primitive()->input_size() == 3 ? 2 : get_primitive()->output_size());
     }
     bool has_second_output() const { return get_output_nums() == 2; }
     bool use_multiple_outputs() const { return get_primitive()->input_size() != 3; }
